code-split: separate the committee fetching from the event fetching
